Move installations data to context
