+++ .changeset/fuzzy-heads-know.md
@@ -1,0 +1,5 @@
+---
+'@giantswarm/backstage-plugin-gs': patch
+---
+
+Moved installations data to context.
