+++ plugins/gs/src/components/installations/InstallationsProvider/index.ts
@@ -1,0 +1,2 @@
+export { InstallationsProvider } from './InstallationsProvider';
+export type { InstallationInfo } from './useInstallationsData';
